requirements.txt
streamlit
xlrd
db-sqlite3
vega-datasets
Pillow
yfinance
yahoo-finance
altair
plotly
pandas
numpy
yahoo_fin
